+++ frontend/src/global_variables.js
@@ -1,4 +1,3 @@
-const backend_url = "https://booking.tgdk.se";
-// const backend_url = "http://localhost:5000";
+const backend_url = "http://localhost:5000";
 
 export default backend_url;